api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

psycopg2-binary
SQLAlchemy
azure-functions
azure-cosmos
google-auth
google-cloud-storage